+++ frontend/src/app/App.jsx
@@ -1,7 +1,7 @@
 import './App.css'
 import {MainHeader} from "../components/MainHeader.jsx";
 import {Route, Routes} from "react-router-dom";
-import RequireAuth from "../components/RequireAuth.jsx";
+import RequireAuth from "../components/common/RequireAuth.jsx";
 import {HomePage} from "../pages/HomePage.jsx";
 import {StoragePage} from "../pages/StoragePage.jsx";