+++ frontend/src/components/Logo.jsx
@@ -1,3 +1,6 @@
+import {FontAwesomeIcon} from "@fortawesome/react-fontawesome";
+import {faCloud} from "@fortawesome/free-regular-svg-icons";
+
 export const Logo = ({onClick}) => {
 
   const handleClick = (e) => {
@@ -7,7 +10,10 @@ export const Logo = ({onClick}) => {
 
   return (
     <>
-      <div className="header-logo" onClick={handleClick}></div>
+      <div className="header-logo btn" onClick={handleClick}>
+        <FontAwesomeIcon icon={faCloud} className="header-cloud-icon" />
+        <span className="header-app-name">My Cloud</span>
+      </div>
     </>
   )
 }