+++ frontend/src/components/MainHeader.jsx
@@ -1,22 +1,48 @@
 import {useAuth} from "../hooks/useAuth.js";
 import {Logo} from "./Logo.jsx";
 import {useNavigate} from "react-router-dom";
-import {UploadFileButton} from "./UploadFileButton.jsx";
+import {UploadFileButton} from "./common/UploadFileButton.jsx";
+import {SearchField} from "./SearchField.jsx";
+import {UserProfileButton} from "./UserProfileButton.jsx";
+import {useState} from "react";
+import RegistrationModal from "./modals/RegistrationModal.jsx";
+import LoginModal from "./modals/LoginModal.jsx";
 
 
 export const MainHeader = () => {
-  const {user} = useAuth()
-  const navigate = useNavigate()
+    const {user, logout} = useAuth()
+    const navigate = useNavigate()
+    const [modalType, setModalType] = useState(null)
 
-  const returnToHome = () => {
-    navigate("/")
-  }
-  return (
-    <>
-      <Logo onClick={returnToHome}/>
-      <div className="header-search-container"></div>
-      {user && <UploadFileButton/>}
-      <div className="header-user-container"></div>
-    </>
-  )
+    const returnToHome = () => {
+        navigate("/")
+    }
+
+    const userIconClick = () => {
+      user ? navigate("/storage") : setModalType('login')
+    }
+
+    const userBtnClick = () => {
+        if (user) {
+            logout()
+        } else {
+            setModalType('login')
+        }
+    }
+
+    return (
+        <>
+            <div className="header-search-area">
+                <Logo onClick={returnToHome}/>
+                <SearchField/>
+            </div>
+            <div className="header-user-area">
+                {user && <UploadFileButton/>}
+                <UserProfileButton iconOnClick={userIconClick} btnOnClick={userBtnClick} btnTitle={user ? 'Выход' : 'Вход'}/>
+            </div>
+
+          {modalType === 'registration' && <RegistrationModal onClose={() => setModalType(null)}/>}
+          {modalType === 'login' && <LoginModal onClose={() => setModalType(null)}/>}
+        </>
+    )
 }