+++ frontend/src/components/common/ContextMenu.jsx
@@ -1,0 +1,9 @@
+export const ContextMenu = ({children, direction, style}) => {
+    return (
+        <>
+            <div className={"context-menu-container " + direction} style={style}>
+                {children}
+            </div>
+        </>
+    )
+}