+++ frontend/src/components/common/CreateItemButton.jsx
@@ -1,0 +1,7 @@
+export const CreateItemButton = () => {
+    return (
+        <>
+            <button className="create-item-btn btn" type="">Создать</button>
+        </>
+    )
+}