+++ frontend/src/components/common/StorageItem.jsx
@@ -1,6 +1,0 @@
-export const StorageItem = () => {
-  return (
-    <>
-    </>
-  )
-}