+++ CHANGELOG.md
@@ -2,14 +2,50 @@
 
 ## Содержание
 
+- [v0.1.0](#v010)
 - [v0.0.1 - Initial version](#v001)
 
 ---
 
 ## Список версий:
 
-<details id="v001" open>
+<details id="v010" open>
+<summary><strong><em>v0.1.0</em></strong></summary>
+
+<p>Начальный интерфейс приложения</p>
+<p>Подробнее:</p>
+
+- <details>
+  <summary><em>Backend</em></summary>
+
+    - Добавлены зависимости в файл requirements.txt
+
+  </details>
+
+- <details>
+  <summary><em>Frontend</em></summary>
+
+    - Реализованы:
+      - UI-компоненты:
+        - `CreateItemButton`, `FileContextMenu`, `FolderContextMenu`, `GridView`, `ListView`, `ListRow`,
+            `SearchField`, `UserProfileButton`
+      - Логика:
+        - Функция `getFileIcon()`
+    - Реализованы два типа отображения файлов
+    - Добавлено тестовое контекстное меню
+    - Добавлена переадресация на страницу хранилища пользователя при нажатии на его иконку, 
+      при условии успешной авторизации на сайте
+    - Различные улучшения отображения компонентов
+
+  </details>
+
+</details>
+
+<details id="v001">
 <summary><strong><em>v0.0.1 - Initial version</em></strong></summary>
+
+<p>Инициализация приложения</p>
+<p>Подробнее:</p>
 
 - <details>
   <summary><em>Backend</em></summary>